+++ main/src/main/java/net/oneandone/stool/setup/UpgradeBuilder.java
@@ -304,7 +304,6 @@ public class UpgradeBuilder {
                     dest.remove("tomcat.editor.userdata");
                 } else {
                     array = new JsonArray();
-                    array.add(new JsonPrimitive(StageConfiguration.NOTIFY_MAINTAINER));
                     array.add(new JsonPrimitive(StageConfiguration.NOTIFY_CREATOR));
                     dest.add("notify", array);
                     dest.add("quota", new JsonPrimitive(10000));