+++ main/src/main/java/net/oneandone/stool/util/Session.java
@@ -173,7 +173,7 @@ public class Session {
             this.users = Users.fromLogin();
         } else {
             this.users = Users.fromLdap(configuration.ldapUrl, configuration.ldapPrincipal, configuration.ldapCredentials,
-                    "ou=users,ou=" + configuration.ldapSso);
+                    "ou=users,ou=" + configuration.ldapUnit);
         }
         this.lockManager = LockManager.create(home.join("run/locks"), user + ":" + command.replace("\n", "\\n"), 30);
         this.lazyPool= null;